Started working on vehicle simulation

+++ src/data/Constants.cc
@@ -1,14 +1,14 @@
 #ifndef __PROJECTS_PSE_TRAFFIC_SIM_SRC_DATA_CONSTANTS_CC_
 #define __PROJECTS_PSE_TRAFFIC_SIM_SRC_DATA_CONSTANTS_CC_
 
-int    VEHICLE_LENGTH           = 4;
+double VEHICLE_LENGTH           = 4;
 double SPEED_MAX                = 16.6;
 double ACCELERATION_MAX         = 1.44; 
 double BRAKE_MAX                = 4.61;
-int    BRAKE_DISTANCE           = 15;
-int    FOLLOW_MIN               = 4;
+double BRAKE_DISTANCE           = 15;
+double FOLLOW_MIN               = 4;
 double SIM_TIME                 = 0.0166;
-int    DECELERATION_DISTANCE    = 50; 
+double DECELERATION_DISTANCE    = 50; 
 double DECELERATION_FACTOR      = 0.4 ;
